move ZoomableContainer to its own file; give credit where credit is due
